// generated/src/app/api-models/claim-hold-determinants.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimHoldDeterminants {
  seqClhldRule: number;
  determinantColumnNo: number;
  determinantTable: string;
  determinant: string;
  operator: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}